requirements.txt
Flask==3.0.0
pandas>=2.3.0
plotly>=5.17.0
Jinja2>=3.1.2
streamlit>=1.28.0
streamlit-authenticator>=0.3.0
# Google Drive API用
google-api-python-client>=2.140.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2 